requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

openai>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0

pydantic>=2.6.0,<3.0.0
slowapi>=0.1.9,<1.0.0

redis>=5.0.1,<6.0.0

pytest>=8.0.0,<9.0.0

prometheus-client>=0.16.0,<1.0.0

python-jose[cryptography]>=3.0.0,<4.0.0